Rev 0.78, Location manual pick implemented
Primary changes to allow the user entry toilets to adjust the location of the toilet by dragging the marker on the map.
Another change is to reverse map the address from the latlng.
Last, User Add Toilet is now split into two activities. First one takes only inputs for toilet onboarding.  For rating we use the same common rating screen.