Rev 0.80 - included Email Authentication with Firebase
Basic implementation for user registration and logging.  Impleted the firebase authentication at teh time of start up
1. email login works
2. Google does not work
3. Phone is not tried yet.
Current User is captued in an object and passed on as an intent parameter to ToiletMaster and UserRating
Exception on reading existing data with no toilet contributor is handled and data is replaced with the current user if the contributor is missing
Rating captures the currentUser.
To do Comments display  to handle missing user information.